suggestion
Kendrick, regarding css it is new for me and probably I am not in the position to criticize. But,  may be it would be better to have a two section footer. I also didn't understand why you have two nav bar one with input and label (for hamburger menu) and the other one inside the mobileHeader.